requirements.txt
matplotlib==3.3.2
plotly==4.14.3
streamlit==0.77.0
pydeck==0.6.1
seaborn==0.11.0
pandas==1.1.3
Pillow==8.1.2
